+++ app/controllers/delorean_test_bed/application_controller.rb
@@ -1,4 +1,4 @@
-module Delorean
+module DeloreanTestBed
   class ApplicationController < ActionController::Base
     protect_from_forgery with: :exception
   end
