--- lib/delorean.rb
@@ -1,5 +1,0 @@
-require "delorean/engine"
-require "delorean/cache"
-
-module Delorean
-end
