--- lib/delorean/version.rb
@@ -1,3 +1,0 @@
-module Delorean
-  VERSION = "0.0.2"
-end
