+++ lib/delorean_test_bed.rb
@@ -1,0 +1,5 @@
+require "delorean_test_bed/engine"
+require "delorean_test_bed/cache"
+
+module Delorean
+end
