+++ lib/delorean_test_bed/version.rb
@@ -1,0 +1,3 @@
+module DeloreanTestBed
+  VERSION = "0.0.3"
+end
