+++ lib/tasks/delorean_test_bed_tasks.rake
@@ -1,4 +1,4 @@
 # desc "Explaining what the task does"
-# task :delorean do
+# task :delorean_test_bed do
 #   # Task goes here
 # end
